QuestionTab class definition

QuestionTab class definition

QuestionTab class definition

Details

Instances are tabs of the questionnaire. This is a wrapper over any ui and server that handles switching to previous/next tab if relevant (if the next_tab/previous_tab argument is NULL there is no button and corresponding server logic).

Public fields

tab_name

name of the tab

tab_title

pretty title to be used in the UI

tab_ui

specific UI function, may take any number of arguments

tab_ui_foot

specific UI function to be used in footer (without arguments)

tab_server

specific server function

tab_number

number assigned automatically based on 'ordered_tabs' vector of names

previous_tab

name of previous tab (optional)

next_tab

name of previous tab (optional)

id

component id in shiny, assigned automatically based on tab number

add_footer

bool argument whether to include the standard footer

add_header

bool argument whether to include the standard header

exposure

optional table of exposures, if present a grid of inputs will be prepared

type

type of institution (to group the inputs from all tabs)

subtype

subtype of tab within a type of institution (to group the inputs from all tabs)

ui_settings

optional list of parameters to pass to ui function

Methods

Public methods

QuestionTab$new()

QuestionTab$server()

QuestionTab$ui()

QuestionTab$clone()

Method new()

the constructor fills the slots with values given it also automatically gets ui, server and foot from the relevant functions (or makes it empty if the function does not exist)

Usage

QuestionTab$new(
  tab_name,
  tab_title,
  previous_tab,
  next_tab,
  add_header = TRUE,
  add_footer = TRUE,
  exposure = NULL,
  type = NULL,
  subtype = NULL,
  ui_settings = list()
)

Arguments

tab_name

name of the tab

tab_title

title of the tab to be used in the UI

previous_tab

name of the previous tab

next_tab

name of the next tab

add_header

whether to add a standard header

add_footer

whether to add a standard footer

exposure

(optional) table of exposure to construct grid of inputs

type

type of issuers for which the inputs are applicable

subtype

unique reference of tab within a type of issuers

ui_settings

(optional) list of arguments to pass to tab_ui

Method server()

Tab server function that combines:

server side of exposure input table (if given in the constructor)

any other server tab_server (if given in the constructor)

possibility of switch to previous/next tab (if applicable), using 'switch_page' function. additionally, a boolean function may be passed to allow going next only conditionally (by default the condition is always true)

Usage

QuestionTab$server(
  input,
  output,
  session,
  switch_page,
  allow_next = function() {     TRUE }
)

Arguments

input

regular shiny parameter

output

regular shiny parameter

session

regular shiny parameter

switch_page

function to be passed that changes the active tab (used in prev/next buttons)

allow_next

(optional) additional condition to be checked before going to next tab

Method ui()

tab UI function that combines:

a common header (unless add_header=FALSE)

title (unless tab_title is NULL)

exposure input table (if exposure table given in the constructore)

any other tab_UI (if given in the constructor)

buttons that switch to previous/next tab (if applicable)

a tab-specific text

a common footer (unless add_footer=FALSE)

Usage

QuestionTab$ui()

Method clone()

The objects of this class are cloneable with this method.

Usage

QuestionTab$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.