Create a single list of settings from simple arguments

Create a single list of settings from simple arguments

get_report_settings(
  content_files,
  output_file,
  md_file,
  file_format,
  report_version,
  dev,
  rep_type,
  inst_type,
  report_sector_selection,
  report_scenario_selection
)

Arguments

content_files

List of necessary global lists with report contents

output_file

Path and filenamename of report to write

md_file

Path and filename of intermediate markdown file

file_format

Currently either "html" or "rtf"

report_version

Integer controlling the version of the code used in report generating functions

dev

Development mode flag (used only to decide if a warning is raised in a lower level function)

rep_type

Either "inst" for institutional report or "sect" for sectoral report

inst_type

Institution type (relevant for institutional report only)

report_sector_selection

Input used to filter report contents

report_scenario_selection

Input used to filter report contents

Value

list of lists