Converts text to HTML and passes to the output

This function should be used on the server side. Corresponding UI side call should be: uiOutput(output_name) All links found in the markdown are modified to be open in the new page (by adding target = "_blank")

include_markdown_text(text, output, output_name, add_new_tab_ref = TRUE)

Arguments

text

text to be converted

output

Shiny output

output_name

Name of slot created in shiny output (to be used in UI part of shiny app)

add_new_tab_ref

if TRUE (the default) all the links will be appended with target="_blank" (to open in the new window)